pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "layra"
description = "Layra is an intelligent CLI tool that scaffolds Python projects using modern best practices and smart presets."
authors = [
    { name = "Andrew Krylov", email = "any@lunte.dev" },
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "typer>=0.15.4",
    "pyyaml>=6.0.2",
    "tomli-w>=0.4.0",
]
dynamic = ['version']

[project.scripts]
layra = "layra.cli:app"

[project.urls]
Repository = "https://github.com/flacy/layra"

[tool.hatch.version]
path = "layra/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["layra"]
